move Order and OrderPartial to testutils package
