+++ fluss-flink/fluss-flink-common/src/test/java/com/alibaba/fluss/flink/sink/serializer/OrderSerializationSchemaTest.java
@@ -16,7 +16,7 @@
 
 package com.alibaba.fluss.flink.sink.serializer;
 
-import com.alibaba.fluss.flink.common.Order;
+import com.alibaba.fluss.flink.source.testutils.Order;
 import com.alibaba.fluss.types.BigIntType;
 import com.alibaba.fluss.types.DataField;
 import com.alibaba.fluss.types.IntType;
