+++ fluss-flink/fluss-flink-common/src/test/java/com/alibaba/fluss/flink/source/FlussSourceITCase.java
@@ -20,13 +20,13 @@ import com.alibaba.fluss.client.table.Table;
 import com.alibaba.fluss.client.table.writer.AppendWriter;
 import com.alibaba.fluss.client.table.writer.UpsertWriter;
 import com.alibaba.fluss.config.ConfigOptions;
-import com.alibaba.fluss.flink.common.Order;
-import com.alibaba.fluss.flink.source.deserializer.OrderPartial;
-import com.alibaba.fluss.flink.source.deserializer.OrderPartialDeserializationSchema;
 import com.alibaba.fluss.flink.source.deserializer.RowDataDeserializationSchema;
 import com.alibaba.fluss.flink.source.enumerator.initializer.OffsetsInitializer;
 import com.alibaba.fluss.flink.source.testutils.FlinkTestBase;
 import com.alibaba.fluss.flink.source.testutils.MockDataUtils;
+import com.alibaba.fluss.flink.source.testutils.Order;
+import com.alibaba.fluss.flink.source.testutils.OrderPartial;
+import com.alibaba.fluss.flink.source.testutils.OrderPartialDeserializationSchema;
 import com.alibaba.fluss.flink.utils.PojoToRowConverter;
 import com.alibaba.fluss.metadata.Schema;
 import com.alibaba.fluss.metadata.TableDescriptor;
